client: pull line.go and switch to injected V1Client.loginFunc
The injection refactoring is the solution of circular import on one side after pulling.
